+++ overlays/kitty.nix
@@ -1,0 +1,5 @@
+self: super: {
+  kitty = super.kitty.overrideAttrs (old: {
+    patches = (old.patches or [ ]) ++ [ ../files/implement-bold_is_bright.patch ];
+  });
+}